+++ backend/test/services/schema.spec.ts
@@ -33,8 +33,8 @@ describe('services > schema', () => {
   test('a schema can be created', async () => {
     SchemaModelModelMock.save.mockResolvedValueOnce(testModelSchema)
     const result = await createSchema(testUser, testModelSchema)
-    expect(SchemaModelModelMock.save).toBeCalledTimes(1)
-    expect(SchemaModelModelMock.deleteOne).not.toBeCalled()
+    expect(SchemaModelModelMock.save).toHaveBeenCalledTimes(1)
+    expect(SchemaModelModelMock.deleteOne).not.toHaveBeenCalled()
     expect(result).toBe(testModelSchema)
   })
 
@@ -46,17 +46,17 @@ describe('services > schema', () => {
     })
 
     const result = () => createSchema(testUser, testModelSchema)
-    await expect(result).rejects.toThrowError(/^You do not have permission to create this schema./)
+    await expect(result).rejects.toThrow(/^You do not have permission to create this schema./)
 
-    expect(SchemaModelModelMock.save).not.toBeCalled()
-    expect(SchemaModelModelMock.deleteOne).not.toBeCalled()
+    expect(SchemaModelModelMock.save).not.toHaveBeenCalled()
+    expect(SchemaModelModelMock.deleteOne).not.toHaveBeenCalled()
   })
 
   test('a schema can be overwritten', async () => {
     SchemaModelModelMock.save.mockResolvedValueOnce(testModelSchema)
     const result = await createSchema(testUser, testModelSchema, true)
-    expect(SchemaModelModelMock.deleteOne).toBeCalledTimes(1)
-    expect(SchemaModelModelMock.save).toBeCalledTimes(1)
+    expect(SchemaModelModelMock.deleteOne).toHaveBeenCalledTimes(1)
+    expect(SchemaModelModelMock.save).toHaveBeenCalledTimes(1)
     expect(result).toBe(testModelSchema)
   })
 
@@ -69,7 +69,7 @@ describe('services > schema', () => {
     SchemaModelModelMock.save.mockRejectedValueOnce(mongoError)
     mockMongoUtils.isMongoServerError.mockReturnValueOnce(true)
 
-    await expect(() => createSchema(testUser, testModelSchema)).rejects.toThrowError(
+    await expect(() => createSchema(testUser, testModelSchema)).rejects.toThrow(
       /^The following is not unique: {"mockKey":"mockValue"}/,
     )
   })
@@ -82,7 +82,7 @@ describe('services > schema', () => {
 
   test('that a schema cannot be retrieved by ID when schema does not exist', async () => {
     SchemaModelModelMock.findOne.mockResolvedValueOnce(undefined)
-    await expect(() => getSchemaById(testModelSchema.id)).rejects.toThrowError(/^The requested schema was not found/)
+    await expect(() => getSchemaById(testModelSchema.id)).rejects.toThrow(/^The requested schema was not found/)
   })
 })