+++ backend/test/services/smtp/smtp.spec.ts
@@ -1,14 +1,13 @@
 import { describe, expect, test, vi } from 'vitest'
 
 import {
-  completeImportNotification,
   dispatchEmailToModelRole,
-  failImportNotification,
   notifyReviewResponseForAccess,
   notifyReviewResponseForRelease,
   requestReviewForAccessRequest,
   requestReviewForRelease,
   startImportNotification,
+  transferCompleteNotification,
 } from '../../../src/services/smtp/smtp.js'
 import { fromEntity } from '../../../src/utils/entity.js'
 import { testReviewResponse } from '../../testUtils/testModels.js'
@@ -144,28 +143,28 @@ describe('services > smtp > smtp', () => {
     vi.spyOn(configMock.smtp, 'enabled', 'get').mockReturnValueOnce(false)
     await requestReviewForRelease('user:user', review, release)
 
-    expect(transporterMock.sendMail).not.toBeCalled()
+    expect(transporterMock.sendMail).not.toHaveBeenCalled()
   })
 
   test('that an Access Request Review email is not sent when disabled in config', async () => {
     vi.spyOn(configMock.smtp, 'enabled', 'get').mockReturnValueOnce(false)
     await requestReviewForAccessRequest('user:user', review, access)
 
-    expect(transporterMock.sendMail).not.toBeCalled()
+    expect(transporterMock.sendMail).not.toHaveBeenCalled()
   })
 
   test('that an email is not sent after a response for a release review if disabled in config', async () => {
     vi.spyOn(configMock.smtp, 'enabled', 'get').mockReturnValueOnce(false)
     await notifyReviewResponseForRelease(testReviewResponse as any, release)
 
-    expect(transporterMock.sendMail).not.toBeCalled()
+    expect(transporterMock.sendMail).not.toHaveBeenCalled()
   })
 
   test('that an email is not sent after a response for a an access request review if disabled in config', async () => {
     vi.spyOn(configMock.smtp, 'enabled', 'get').mockReturnValueOnce(false)
     await notifyReviewResponseForAccess(testReviewResponse as any, access)
 
-    expect(transporterMock.sendMail).not.toBeCalled()
+    expect(transporterMock.sendMail).not.toHaveBeenCalled()
   })
 
   test('that an email is sent for Release Reviews', async () => {
@@ -199,13 +198,20 @@ describe('services > smtp > smtp', () => {
   })
 
   test('that an email is sent after an import has complete', async () => {
-    await completeImportNotification('modelId', ['fileabc'], ['alpine:latest'])
+    await transferCompleteNotification('modelId', false, {
+      'Successful Files': ['fileabc'],
+      'Successful Images': ['alpine:latest'],
+    })
 
     expect(transporterMock.sendMail.mock.calls.at(0)).toMatchSnapshot()
   })
 
   test('that an email is sent after an import has failed', async () => {
-    await failImportNotification('modelId', ['fileabcd'], ['alpine:old'], ['fileabc'], ['alpine:latest'])
+    await transferCompleteNotification('modelId', true, {
+      'Failed Images': ['alpine:old'],
+      'Successful Files': ['fileabc'],
+      'Successful Images': ['alpine:latest'],
+    })
 
     expect(transporterMock.sendMail.mock.calls.at(0)).toMatchSnapshot()
   })
@@ -252,6 +258,6 @@ describe('services > smtp > smtp', () => {
     transporterMock.sendMail.mockRejectedValueOnce('Failed to send email')
 
     const result: Promise<void> = requestReviewForRelease('user:user', review, release)
-    await expect(result).rejects.toThrowError(`Unable to send email`)
+    await expect(result).rejects.toThrow(`Unable to send email`)
   })
 })